Do not require `renumber_vertex_indices()` for VertexIndexGraph concept
This requirement was previously marked as relaxed and seems not essential
for the concept. Indeed one can access and use the index without updating
it.

With this change adjacency_list<vecS> implements the concept, and this
concept can be used to check if vertices are indexed.

The only 2 places in the codebase, where this concept is referred have no
mention of `renumber_vertex_indices`:

detail/index.hpp

tiernan_all_cycles.hpp

The requirement was initially introduces together with `directed_graph`
and `undirected_graph`, which both provide this function.

I'm not completely sure about this change, but it seems to go inline with
the BGL principles.